# [Querier] optimize prom query