… server/querier/app/prometheus/service/analyzer.go
… server/querier/app/prometheus/service/functions.go
… server/querier/app/prometheus/service/promql.go
… server/querier/app/prometheus/service/service.go