<!--

Thank you for contributing to DeepFlow!
Please read this template before submitting pull requests.
Texts surrounded by `<` and `>` should be replaced accordingly.
Put an `x` in `[ ]` to mark the item as checked. `[x]`

-->

### This PR is for:
- Server
<!--
One or more of:
- Agent
- CLI
- Server
- Message
- Libs
- Documents
- Workflow
-->

### Improves the performance of promql query
#### Added benchmark
- base on #5462 , add more optimize
- only v6.4+: add `rate` implement for operator-offloading
  - before: when use `rate` to query huge data like `rate(m[3d])` or `rate(m[7d]`, it requires 3d/7d data for calculation.
  - after: consider the `step` in range query, when use `step` for calculation, it's downsampling for data load. So when use `range query` with `step`, NOW we will load the specific calculate data points into memory, lower the data sets for loading.

#### Benchmark result
- base on #5462 , add more tests with `operator-offloading=true`
query range: `1h`, step: `10m`
promql: 
```text
((sum by (cluster) (rate(apiserver_request_duration_seconds_count{job="apiserver",verb=~"LIST|GET"}[1h])) - ((sum by (cluster) (rate(apiserver_request_duration_seconds_bucket{job="apiserver",le="1",scope=~"resource|",verb=~"LIST|GET"}[1h])) or vector(0)) + sum by (cluster) (rate(apiserver_request_duration_seconds_bucket{job="apiserver",le="5",scope="namespace",verb=~"LIST|GET"}[1h])) + sum by (cluster) (rate(apiserver_request_duration_seconds_bucket{job="apiserver",le="30",scope="cluster",verb=~"LIST|GET"}[1h])))) + sum by (cluster) (rate(apiserver_request_total{code=~"5..",job="apiserver",verb=~"LIST|GET"}[1h]))) / sum by (cluster) (rate(apiserver_request_total{job="apiserver",verb=~"LIST|GET"}[1h]))
```

1. use cache ( with mock user 20, ~20qps)
- before test results: ref to #5462 

- after:

|         | avg(ms) | max(ms) | min(ms) | p90 |
|---------|---------|---------|---------|-----|
| range   | 53     | 10126   | 4 | 17 |
| instant | 11     | 4615    | 4 | 15 |

![after-offloading-cache-20](https://github.com/deepflowio/deepflow/assets/45836837/dfd978b2-8936-4d01-93ee-b2978bf4150a)

2. not using cache 
- before test results: ref to #5462 (when query without cache, previous test results could only handle 5 user qps)

- after: (with mock user 5, ~5qps) **compare to previous tests**

| | avg(ms) | max(ms) | min(ms) | p90 |
|---------|---------|---------|---------|-----|
| range   | 6525 | 10395 | 3282 | 8252 |
| instant | 5522 | 8568 | 3000 | 6775 |

![after-offloading-nocache-5](https://github.com/deepflowio/deepflow/assets/45836837/5a16a0a0-70f8-41a4-be23-5879402b5bc5)

---
- additionally, add 4 times of qps for test:

- with mock user 20, ~20qps

| | avg(ms) | max(ms) | min(ms) | p90 |
|---------|---------|---------|---------|-----|
| range   | 15073     | 24327   | 8366      | 19424 |
| instant | 13587     | 23214    | 7976      | 16624 |

![after-offloading-nocache-20](https://github.com/deepflowio/deepflow/assets/45836837/c20d4c87-fc84-4bd2-8217-157ffd810489)

3. conclusion:
` lower avg 40~50% query duration, add about 3~4 times throughput during query`